collector/requirements.txt
requests
elasticsearch-dsl
tweepy
fire
pyyaml
